requirements.txt
Django>=2.2,<3.0
django-allauth>=0.44.0
django-otp>=1.0.0
django-two-factor-auth>=1.13.0
python-decouple>=3.6
django-crispy-forms>=1.8.0,<1.12.0
pillow>=7.1.0,<8.0.0
djangorestframework>=3.11.0
djangorestframework-simplejwt>=4.4.0
pycryptodome>=3.9.0 